pyproject.toml
[tool.poetry]
name = "mykobo-py"
version = "0.1.0"
description = "Python library to interact with the MyKobo APIs"
authors = ["Kwabena Aning <kwabena.aning@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dataclasses-json = "^0.6.7"
requests = "^2.32.3"
pyjwt = "^2.10.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
requests-mock = "^1.12.1"
python-dotenv = "^1.0.1"
python-semantic-release = "^9.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
